Use the ViewHelper to resize the main dialog
